Added `DataHandler` class
Contacts [x]
Photos   [ ]
Diary    [ ]
